Moved message alert to sparkle
